requirements.txt
torch~=2.8.0
aioboto3~=14.1.0
hydra-core~=1.3.2
# install nmslib from GitHub (instead of PyPI)
git+https://github.com/nmslib/nmslib/#subdirectory=python_bindings
numpy~=1.26.4
omegaconf~=2.3.0
pandas~=1.5.3
pytorch-lightning==2.5.3
pytorch-metric-learning==2.3.0
ranx~=0.3.20
retriv~=0.2.3
scikit-learn~=1.3.2
scipy~=1.11.4
seaborn~=0.13.2
tensorboard~=2.15.2
torchmetrics==1.2.1
tqdm~=4.66.6
transformers~=4.35.2
sentence-transformers~=3.0.1
wandb~=0.16.6
numba~=0.58.1
aiohttp~=3.9.5
rich~=13.7.1
boto3~=1.37.1
